Append to boot-args instead of replace
When adding -no_compat_check to boot-args, do not replace other flags in boot-args that the user may be depending on. For example, a MacPro3,1 with 64 GB of RAM requires maxmem=63488 so that the installer does not take forever. Other flags, such as -v, may be useful for debugging.
This code is analogous to the code that adds -no_compat_check to com.apple.Boot.plist